Adds version of teacher notebook that's runnable in Mybinder
The original version of the teacher notebook expects all the data files
in the local directory (i.e., the same that the notebook is in), which
is where they will have been downloaded and/or moved to by the learners.
This won't work on Mybinder, because the data are in ../data.

This copy corrects the data file paths accordingly, and cuts out the
empty cells and explanations for student exercises.